--- components/basecomponents/MemoryGame/ImageSuspense.module.css
@@ -1,3 +1,0 @@
-.hidden {
-  opacity: 0.5;
-}
