+++ components/basecomponents/MemoryGame/Themes/MemoryGameXmasTheme.module.css
@@ -9,6 +9,23 @@
   margin-bottom: 1.25rem;
 }
 
+.buttonWrapper img, .cell img {
+  position: absolute;  
+  inset: 0px;
+  box-sizing: border-box;
+  padding: 0px;
+  border: none;
+  margin: auto;
+  display: block;
+  width: 0px;
+  height: 0px;
+  min-width: 100%;
+  max-width: 100%;
+  min-height: 100%;
+  max-height: 100%;
+  object-fit: cover;
+}
+
 .newGameButton {
   position: absolute;
   background-color: transparent;
